requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
docstring-parser==0.14.1
executing==2.0.1
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
py-algorand-sdk==2.9.1
pycparser==2.22
pycryptodomex==3.23.0
PyNaCl==1.5.0
pyteal==0.27.0
python-dotenv==1.1.0
requests==2.32.4
semantic-version==2.10.0
tabulate==0.9.0
urllib3==2.4.0
Werkzeug==3.1.3
